@Shivneel Rattan WARNING: This is a complete change of how the processes work. Please understand the changes before working on the code. There is now three endpoints, updatePricebook, getPageCount, and getPage.
The first endpoint that gets triggered is the updatePricebook. When this is called, it will take an email, and convert into a json file stored in Output, with the label of the file having the account number and page.

The second endpoint in the process is getPageCount. When this is called, it gets the number of pages matching the account number / pricebook number/id. This will be how many requests will be made to the getPage endpoint to transfer data.

The third endpoint in the process in the process is getPage. When this is called, it looks into the Output folder, and finds the first json file that has the account number and page number in the correct format. Once found, it will return this to the caller.